refactor: 팀원 추가하는 기능 공통 컴포넌트로 분리
