+++ idearly/src/pages/TeamMatchingPage/TeamMatchingPage.styles.tsx
@@ -1,5 +1,5 @@
 import styled from "@emotion/styled";
-import { Button, Card, CardBody, CardFooter, CardHeader, HStack, Heading, Tag, TagLabel, Text } from "@chakra-ui/react";
+import { Card, CardBody, CardFooter, CardHeader, Heading, Text } from "@chakra-ui/react";
 
 export const TeamMathingWrapper = styled.div`
   width: 100%;
@@ -30,30 +30,7 @@ export const CardFooterSection = styled(CardFooter)`
   justify-content: center;
 `;
 
-export const HStackWrapper = styled(HStack)`
-  margin-bottom: 0.7rem;
-`;
-
-export const TagLabelName = styled(TagLabel)`
-  font-size: 0.8rem;
-`;
-
-export const TagLabeEmail = styled(TagLabel)`
-  font-size: 0.6rem;
-`;
-
-export const TagWrapper = styled(Tag)`
-  background-color: #CBD5E0;
-  color: #4A5568;
-`;
-
 export const MiniTitle = styled(Text)`
   font-weight: bold;
   margin-bottom: 8px;
-`
-
-export const ShowUserTag = styled(Tag)`
-  width: 100%;
-  margin-top: 10px;
-  padding: 10px;
 `
